Load bigfoot only if needed
It appears that sometimes bigfoot has some unwanted side effects, like
shifting the footnote to the next page.

This change reverts to the standard footnotes if the there is no need
to use the secondary footnotes, avoiding to load bigfoot.

Also if the template appears custom (and not updated), keep the
existing behavior to avoid breakages.